refactor(agents): Merge V2 agents view into main agents page
Updates the main agents page with the V2 implementation featuring:
- Enhanced card-based grid layout
- Last updated timestamp display
- Improved visual styling with gradients
- Direct links to canary view

Removes the separate agents-v2 route and consolidates the functionality into
the main /agents path for a more streamlined experience.
